Added authentication manager for Google IAP
