+++ src/main/java/org/ohdsi/webapi/shiro/filters/AtlasAuthFilter.java
@@ -1,0 +1,9 @@
+package org.ohdsi.webapi.shiro.filters;
+
+import org.apache.shiro.web.filter.authc.AuthenticatingFilter;
+
+/**
+ * Filter to check for a user, who has already logged in using external service
+ */
+public abstract class AtlasAuthFilter extends AuthenticatingFilter {
+}
