+++ src/main/java/org/ohdsi/webapi/shiro/management/Security.java
@@ -4,13 +4,11 @@ import java.util.Map;
 import java.util.Set;
 import org.apache.shiro.authc.Authenticator;
 import org.apache.shiro.realm.Realm;
-import org.springframework.stereotype.Component;
 
 /**
  *
  * @author gennadiy.anisimov
  */
-@Component
 public abstract class Security {
 
   public static String PROFILE_VIEW_DATES_PERMISSION = "*:person:*:get:dates";
