+++ src/main/resources/application.properties
@@ -91,7 +91,7 @@ csrf.disable=true
 
 sparql.endpoint=http://virtuoso.ohdsi.org:8890/sparql?default-graph-uri=&query=
 
-security.enabled=${security.enabled}
+security.provider=${security.provider}
 security.cors.enabled=${security.cors.enabled}
 security.token.expiration=${security.token.expiration}
 security.origin=${security.origin}
@@ -119,6 +119,9 @@ security.ad.searchBase=${security.ad.searchBase}
 security.ad.principalSuffix=${security.ad.principalSuffix}
 security.ad.system.username=${security.ad.system.username}
 security.ad.system.password=${security.ad.system.password}
+
+security.googleIap.cloudProjectId=${ssecurity.googleIap.cloudProjectId}
+security.googleIap.backendServiceId=${security.googleIap.backendServiceId}
 
 security.kerberos.spn=${security.kerberos.spn}
 security.kerberos.keytabPath=${security.kerberos.keytabPath}
